backend/requirements.txt
beautifulsoup4==4.8.2
fastapi==0.115.12
Markdown==3.7
matplotlib==3.8.3
mistralai==1.6.0
openai==1.72.0
pandas==2.0.3
passlib==1.7.4
pydantic>=2.10.0  # Use a newer version compatible with pydantic-core 2.27.2
pydantic_settings==2.8.1
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.1.0
python_docx==1.1.2
python_jose==3.4.0
Requests==2.32.3
supabase==2.15.0
uvicorn==0.34.0